requirements.txt
anyio==3.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
dateparser==1.1.7
ecdsa==0.18.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
idna==3.4
mysql-connector-python==8.0.32
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.15
youtube-comment-downloader==0.1.67
youtube-search-python==1.6.6
